use white vertical divider; don't change background
The previous way of doing it made the divider invisible in the GUI and just
barely visible in the terminal. Apparently it isn't possible in the GUI to
treat the divider as actual characters, as gVim does; my intent was to mirror
my Vim/gVim settings, which used light gray pipe characters on a dark gray
background (different from the overall background).